third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/v0/get-destination-insertion-points-re-distribution-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests for getDestinationInsertionPoints() which involves re-distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
